+++ HyperMetro/task/src/metro/command/Action.java
@@ -1,4 +1,4 @@
-package metro.action;
+package metro.command;
 
 import metro.model.MetroMap;
