--- HyperMetro/task/src/metro/command/Command.java
@@ -1,8 +1,0 @@
-package metro.command;
-
-import java.util.List;
-import java.util.function.Consumer;
-
-public interface Command extends Consumer<List<String>> {
-    String name();
-}
