--- HyperMetro/task/src/metro/command/Remove.java
@@ -1,12 +1,0 @@
-package metro.command;
-
-import metro.entity.Line;
-import metro.entity.Metro;
-import metro.ui.UserInterface;
-
-public class Remove extends ManageStation {
-    public Remove(final Metro metro, final UserInterface ui) {
-        super(metro, ui, Line::remove);
-    }
-
-}
