--- HyperMetro/task/src/metro/entity/Line.java
@@ -1,58 +1,0 @@
-package metro.entity;
-
-import java.util.*;
-
-public class Line extends NamedEntity implements Iterable<Station> {
-    private static final List<String> DEPOT = Collections.emptyList();
-    private final LinkedList<Station> stations = new LinkedList<>();
-
-    public Line(final String lineName) {
-        super(lineName);
-    }
-
-    @Override
-    public Iterator<Station> iterator() {
-        return stations.iterator();
-    }
-
-    public Optional<Station> getByName(String stationName) {
-        return stations.stream()
-                .filter(station -> station.isNameEquals(stationName))
-                .findAny();
-    }
-
-    public void append(final String stationName) {
-        append(new JsonStation(stationName));
-    }
-
-    public void append(final JsonStation station) {
-        final List<String> prevStations;
-
-        if (stations.isEmpty()) {
-            prevStations = DEPOT;
-        } else {
-            stations.getLast().setNextStations(List.of(station.getName()));
-            prevStations = List.of(stations.getLast().name());
-        }
-        stations.add(new Station(station, prevStations, DEPOT));
-    }
-
-    public void addHead(final String stationName) {
-        addHead(new JsonStation(stationName));
-    }
-
-    public void addHead(final JsonStation station) {
-        final List<String> nextStations;
-        if (stations.isEmpty()) {
-            nextStations = DEPOT;
-        } else {
-            stations.getFirst().setPreviousStations(List.of(station.getName()));
-            nextStations = List.of(stations.getFirst().name());
-        }
-        stations.addFirst(new Station(station, DEPOT, nextStations));
-    }
-
-    public void remove(final String stationName) {
-        throw new UnsupportedOperationException("operation is not implemented");
-    }
-}
