--- HyperMetro/task/src/metro/json/Metro.java
@@ -1,15 +1,0 @@
-package metro.json;
-
-import java.util.Map;
-
-public class Metro {
-    Map<String, Object> lines;
-
-    public Map<String, Object> getLines() {
-        return lines;
-    }
-
-    public void setLines(final Map<String, Object> lines) {
-        this.lines = lines;
-    }
-}
